+++ test/avocado/run-tests
@@ -27,10 +27,10 @@ import sys
 # the directory structure breaks on the second run with .pyc files
 sys.dont_write_bytecode = True
 
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-sys.path.append(os.path.abspath(os.path.dirname(os.path.realpath(__file__))))
+sys.path.append(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
+from verify import parent
 
-from common import testvm
+import testvm
 
 machine_test_dir = "/tmp/avocado_tests"
