+++ test/containers/check-kubernetes
@@ -21,12 +21,12 @@
 import os
 import sys
 
-base_dir = os.path.dirname(os.path.realpath(__file__))
-test_dir = os.path.abspath(os.path.join(base_dir, ".."))
+test_dir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 sys.path.append(test_dir)
 
-from common.testlib import *
-from common.testvm import VirtMachine
+from verify import parent
+from testlib import *
+from testvm import VirtMachine
 from verify import kubelib
 
 class KubernetesContainerMachine(VirtMachine):
