+++ test/verify/check-services
@@ -20,7 +20,7 @@
 
 import parent
 from testlib import *
-from common import testvm
+import testvm
 
 class TestServices(MachineCase):
     def setUp(self):
