+++ test/vm-run
@@ -22,10 +22,8 @@ import os
 import subprocess
 import sys
 
-from common import testvm
-
-TEST = os.path.dirname(__file__)
-BOTS = os.path.normpath(os.path.join(TEST, "..", "bots"))
+from verify import parent
+import testvm
 
 NETWORK_SCRIPT="""
     set -ex
@@ -92,7 +90,7 @@ try:
         kubeconfig = os.path.join(os.path.expanduser("~"), ".kube", "config")
         if not os.path.lexists(kubeconfig):
             d = os.path.dirname(kubeconfig)
-            src = os.path.abspath(os.path.join(BOTS, "images", "files", "openshift.kubeconfig"))
+            src = os.path.abspath(os.path.join(parent.BOTS_DIR, "images", "files", "openshift.kubeconfig"))
             if not os.path.exists(d):
                 os.makedirs(d)
             sys.stderr.write("image-run: linking kubeconfig into ~/.kube/config\n")
@@ -101,7 +99,7 @@ try:
     # Check that image is downloaded
     if not os.path.exists(machine.image_file):
         try:
-            ret = subprocess.call([ os.path.join(BOTS, "image-download"), args.image])
+            ret = subprocess.call([ os.path.join(parent.BOTS_DIR, "image-download"), args.image])
         except OSError as ex:
             if ex.errno != errno.ENOENT:
                 raise
